Include userId in admin event
The userId that was included in admin events was different than the
userId included in regular events, as the userId for admin events was
the id of the admin performing a change. This can be confusing since
both events are treated as the same in PMP, meaning that unless the
developer has a good understanding of this event listener, the userId
will be inconsistent between events.

This changes the userId in admin events to be the id of the user that
the change happened for, and introduces a new field for the id of the
admin that performed the change. This hopefully makes the userId
consistent for developers of PMP and makes it possible to distinguish
admin event from regular events should it be necessary.